auth small changes
